+++ src/parser/MainParser.ts
@@ -9,41 +9,46 @@ import MacroParser from './macro/MacroParser';
 import SjasmplusFakeInstructionParser from './sjasmplus/SjasmplusFakeInstructionParser';
 import SjasmplusRegisterListInstructionParser from './sjasmplus/SjasmplusRegisterListInstructionParser';
 import Z80InstructionParser from './z80/Z80InstructionParser';
+import SourceCodeLine from '../model/SourceCodeLine';
+import SourceCodePart from '../model/SourceCodePart';
+import SubroutineTimingHintDecorator from '../timing/SubroutineTimingHintDecorator';
 
 export default class MainParser {
 
     // Configuration
     private platformConfiguration: string;
-    private sjasmplus: boolean;
+    private sjasmplusConfiguration: boolean;
     private syntaxLabelConfiguration: string;
     private syntaxLineSeparatorConfiguration: string;
     private syntaxRepeatConfiguration: string;
+    private timingsHintsConfiguration: boolean;
 
     constructor() {
 
         // Saves configuration
         const configuration = workspace.getConfiguration("z80-asm-meter");
         this.platformConfiguration = configuration.get("platform", "z80");
-        this.sjasmplus = configuration.get("sjasmplus", false);
+        this.sjasmplusConfiguration = configuration.get("sjasmplus", false);
         this.syntaxLabelConfiguration = configuration.get("syntax.label", "default");
         this.syntaxLineSeparatorConfiguration = configuration.get("syntax.lineSeparator", "none");
         this.syntaxRepeatConfiguration = configuration.get("syntax.repeat", "none");
+        this.timingsHintsConfiguration = configuration.get("timings.hints", false);
     }
 
-    parse(sourceCode: string | undefined): MeterableCollection {
+    parse(rawSourceCode: string | undefined): MeterableCollection {
 
         const meterables = new MeterableCollection();
 
         // (sanity checks)
-        if (!sourceCode) {
+        if (!rawSourceCode) {
             return meterables;
         }
-        const rawLines = sourceCode.split(/[\r\n]+/);
-        if (rawLines.length === 0) {
+        const rawSourceCodeLines = rawSourceCode.split(/[\r\n]+/);
+        if (rawSourceCodeLines.length === 0) {
             return meterables;
         }
-        if (rawLines[rawLines.length - 1].trim() === "") {
-            rawLines.pop(); // (removes possible spurious empty line at the end of the selection)
+        if (rawSourceCodeLines[rawSourceCodeLines.length - 1].trim() === "") {
+            rawSourceCodeLines.pop(); // (removes possible spurious empty line at the end of the selection)
         }
 
         // Determines syntax
@@ -56,52 +61,49 @@ export default class MainParser {
                 : undefined;
 
         // Extracts the instructions for every line
-        rawLines.forEach(rawLine => {
-            const rawInstructionCandidates =
-                    this.extractRawInstructionCandidatesFrom(rawLine, labelRegExp, lineSeparator);
+        rawSourceCodeLines.forEach(rawSourceCodeLine => {
+            const sourceCodeLines =
+                    this.extractSourceCodeLinesFrom(rawSourceCodeLine, labelRegExp, lineSeparator);
 
             // Parses the instructions
-            rawInstructionCandidates?.forEach((rawInstructionCandidate: string | undefined) => {
-                meterables.add(this.parseRawInstructionCandidate(rawInstructionCandidate));
+            sourceCodeLines.getParts().forEach((sourceCodePart: SourceCodePart) => {
+                meterables.add(this.parseSourceCodePart(sourceCodePart));
             });
         });
 
         return meterables;
     }
 
-    private extractRawInstructionCandidatesFrom(
-        rawLine: string, labelRegExp: RegExp,
-        lineSeparator: string | undefined): string[] | undefined {
+    private extractSourceCodeLinesFrom(
+        rawSourceCodeLine: string, labelRegExp: RegExp,
+        lineSeparator: string | undefined): SourceCodeLine {
 
         // Removes surrounding label and/or whitespace
-        const rawParts = rawLine.replace(labelRegExp, "").trim();
+        const cleanRawSourceCodeLine = rawSourceCodeLine.replace(labelRegExp, "").trim();
 
-        // For every part of the line
-        const rawInstructionCandidates: string[] = [];
-        normalizeAndSplitQuotesAware(rawParts, lineSeparator).forEach(part => {
-            if (part.length !== 0) {
-                rawInstructionCandidates.push(part);
-            }
-        });
-
-        return rawInstructionCandidates.length === 0 ? undefined : rawInstructionCandidates;
+        // Splits the line and extracts trailing comment
+        return normalizeAndSplitQuotesAware(cleanRawSourceCodeLine, lineSeparator);
     }
 
-    private parseRawInstructionCandidate(
-            rawInstructionCandidate: string | undefined): Meterable | undefined {
+    private parseSourceCodePart(
+            sourceCodePart: SourceCodePart): Meterable | undefined {
 
-        // (sanity check)
-        if (!rawInstructionCandidate) {
-            return undefined;
+        const rawPart = sourceCodePart.getPart();
+        const rawInstruction = this.extractRawInstruction(rawPart);
+        var meterable = this.parseRawInstruction(rawInstruction);
+
+        // Tries to parse timing hints from the comment
+        if (this.timingsHintsConfiguration) {
+            const rawComment = sourceCodePart.getComment();
+            meterable = SubroutineTimingHintDecorator.of(meterable, rawComment);
         }
 
         // Tries to parse the optional repeat pseudo-op
-        const repeatCount = this.extractRepeatCount(rawInstructionCandidate);
-        const rawInstruction = this.extractRawInstruction(rawInstructionCandidate);
-        return MeterableRepetition.of(this.parseRawInstruction(rawInstruction), repeatCount);
+        const repeatCount = this.extractRepeatCount(rawPart);
+        return MeterableRepetition.of(meterable, repeatCount);
     }
 
-    private extractRepeatCount(rawInstructionCandidate: string): number {
+    private extractRepeatCount(s: string): number {
 
         // Determines syntax
         const repeatRegExp =
@@ -113,7 +115,7 @@ export default class MainParser {
         }
 
         // Tries to parse repeat pseudo-op
-        const matches = repeatRegExp.exec(rawInstructionCandidate);
+        const matches = repeatRegExp.exec(s);
         const hasRepeatCount = matches && matches.length >= 1 && matches[1];
         if (!hasRepeatCount) {
             return 1;
@@ -122,7 +124,7 @@ export default class MainParser {
         return repeatCountCandidate && repeatCountCandidate > 0 ? repeatCountCandidate : 1;
     }
 
-    private extractRawInstruction(rawInstructionCandidate: string): string {
+    private extractRawInstruction(s: string): string {
 
         // Determines syntax
         const repeatRegExp =
@@ -130,16 +132,16 @@ export default class MainParser {
             : this.syntaxRepeatConfiguration === "dot" ? /^(?:\.(\S+)\s)(.+)$/
             : undefined;
         if (!repeatRegExp) {
-            return rawInstructionCandidate;
+            return s;
         }
 
         // Tries to parse beyond the repeat pseudo-op
-        const matches = repeatRegExp.exec(rawInstructionCandidate);
+        const matches = repeatRegExp.exec(s);
         const hasRepeatInstruction = matches && matches.length >= 2 && matches[2];
-        return hasRepeatInstruction ? matches[2] : rawInstructionCandidate;
+        return hasRepeatInstruction ? matches[2] : s;
     }
 
-    private parseRawInstruction(rawInstruction: string): Meterable | undefined {
+    private parseRawInstruction(s: string): Meterable | undefined {
 
         // Determines instruction sets
         const instructionSets =
@@ -147,33 +149,33 @@ export default class MainParser {
             : [ "S" ];
 
         // Tries to parse Z80 instructions
-        const z80Instruction = Z80InstructionParser.instance.parseInstruction(rawInstruction, instructionSets);
+        const z80Instruction = Z80InstructionParser.instance.parseInstruction(s, instructionSets);
         if (!!z80Instruction) {
             return z80Instruction;
         }
 
         // Tries to parse sjasmplus alternative syntax and fake instructions
-        if (this.sjasmplus) {
+        if (this.sjasmplusConfiguration) {
             const sjasmplusFakeInstruction =
-                    SjasmplusFakeInstructionParser.instance.parse(rawInstruction, instructionSets);
+                    SjasmplusFakeInstructionParser.instance.parse(s, instructionSets);
             if (!!sjasmplusFakeInstruction) {
                 return sjasmplusFakeInstruction;
             }
             const sjasmplusRegisterListInstruction =
-                    SjasmplusRegisterListInstructionParser.instance.parse(rawInstruction, instructionSets);
+                    SjasmplusRegisterListInstructionParser.instance.parse(s, instructionSets);
             if (sjasmplusRegisterListInstruction) {
                 return sjasmplusRegisterListInstruction;
             }
         }
 
         // Tries to parse user-defined macro
-        const macro = MacroParser.instance.parse(rawInstruction, instructionSets);
+        const macro = MacroParser.instance.parse(s, instructionSets);
         if (!!macro) {
             return macro;
         }
 
         // Tries to parse assembly directives
-        const assemblyDirective = AssemblyDirectiveParser.instance.parse(rawInstruction);
+        const assemblyDirective = AssemblyDirectiveParser.instance.parse(s);
         if (!!assemblyDirective) {
             return assemblyDirective;
         }
@@ -183,4 +185,3 @@ export default class MainParser {
     }
 
 }
-